+++ tests/spec/routeTypes/schema.ts
@@ -1120,7 +1120,7 @@ export namespace gists {
   */
   export namespace CommentsCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = commentBody;
     export type ResponseBody = comment;
   }
 
@@ -1356,7 +1356,7 @@ export namespace notifications {
   */
   export namespace NotificationsUpdate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = notificationMarkRead;
     export type ResponseBody = any;
   }
 
@@ -1560,7 +1560,7 @@ export namespace orgs {
   */
   export namespace ReposCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = postRepo;
     export type ResponseBody = repos;
   }
 
@@ -1770,7 +1770,7 @@ export namespace repos {
   */
   export namespace CommentsPartialUpdate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = commentBody;
     export type ResponseBody = commitComment;
   }
 
@@ -2161,7 +2161,7 @@ export namespace repos {
   */
   export namespace HooksCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = hookBody;
     export type ResponseBody = hook;
   }
 
@@ -2196,7 +2196,7 @@ export namespace repos {
   */
   export namespace HooksPartialUpdate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = hookBody;
     export type ResponseBody = hook;
   }
 
@@ -2229,7 +2229,7 @@ export namespace repos {
   */
   export namespace IssuesCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = issue;
     export type ResponseBody = issue;
   }
 
@@ -2275,7 +2275,7 @@ export namespace repos {
   */
   export namespace IssuesCommentsPartialUpdate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = commentBody;
     export type ResponseBody = issuesComment;
   }
 
@@ -2323,7 +2323,7 @@ export namespace repos {
   */
   export namespace IssuesPartialUpdate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = issue;
     export type ResponseBody = issue;
   }
 
@@ -2347,7 +2347,7 @@ export namespace repos {
   */
   export namespace IssuesCommentsCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = commentBody;
     export type ResponseBody = issuesComment;
   }
 
@@ -2393,7 +2393,7 @@ export namespace repos {
   */
   export namespace IssuesLabelsCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = emailsPost;
     export type ResponseBody = label;
   }
 
@@ -2404,7 +2404,7 @@ export namespace repos {
   */
   export namespace IssuesLabelsUpdate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = emailsPost;
     export type ResponseBody = label;
   }
 
@@ -2439,7 +2439,7 @@ export namespace repos {
   */
   export namespace KeysCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = UserKeysPost;
     export type ResponseBody = UserKeysKeyId;
   }
 
@@ -2485,7 +2485,7 @@ export namespace repos {
   */
   export namespace LabelsCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = emailsPost;
     export type ResponseBody = label;
   }
 
@@ -2520,7 +2520,7 @@ export namespace repos {
   */
   export namespace LabelsPartialUpdate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = emailsPost;
     export type ResponseBody = label;
   }
 
@@ -2564,7 +2564,7 @@ export namespace repos {
   */
   export namespace MilestonesCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = milestoneUpdate;
     export type ResponseBody = milestone;
   }
 
@@ -2599,7 +2599,7 @@ export namespace repos {
   */
   export namespace MilestonesPartialUpdate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = milestoneUpdate;
     export type ResponseBody = milestone;
   }
 
@@ -2632,7 +2632,7 @@ export namespace repos {
   */
   export namespace NotificationsUpdate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = notificationMarkRead;
     export type ResponseBody = any;
   }
 
@@ -2700,7 +2700,7 @@ export namespace repos {
   */
   export namespace PullsCommentsPartialUpdate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = commentBody;
     export type ResponseBody = pullsComment;
   }
 
@@ -2825,7 +2825,7 @@ export namespace repos {
   */
   export namespace ReleasesCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = ReleaseCreate;
     export type ResponseBody = release;
   }
 
@@ -2893,7 +2893,7 @@ export namespace repos {
   */
   export namespace ReleasesPartialUpdate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = ReleaseCreate;
     export type ResponseBody = release;
   }
 
@@ -3360,7 +3360,7 @@ export namespace user {
   */
   export namespace EmailsCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = emailsPost;
     export type ResponseBody = any;
   }
 
@@ -3448,7 +3448,7 @@ export namespace user {
   */
   export namespace KeysCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = UserKeysPost;
     export type ResponseBody = UserKeysKeyId;
   }
 
@@ -3503,7 +3503,7 @@ export namespace user {
   */
   export namespace ReposCreate {
     export type RequestQuery = {};
-    export type RequestBody = any;
+    export type RequestBody = postRepo;
     export type ResponseBody = repos;
   }
 
@@ -3965,7 +3965,7 @@ export class Api<SecurityDataType> {
     * @request POST:/gists/{id}/comments
     * @description Create a commen
     */
-    commentsCreate: (id: number, body: any, params?: RequestParams) =>
+    commentsCreate: (id: number, body: commentBody, params?: RequestParams) =>
       this.request<comment>(`/gists/${id}/comments`, "POST", params, body),
 
 
@@ -4161,7 +4161,7 @@ export class Api<SecurityDataType> {
     * @request PUT:/notifications
     * @description Mark as read.. Marking a notification as "read" removes it from the default view on GitHub.com.. 
     */
-    notificationsUpdate: (body: any, params?: RequestParams) =>
+    notificationsUpdate: (body: notificationMarkRead, params?: RequestParams) =>
       this.request<any>(`/notifications`, "PUT", params, body),
 
 
@@ -4329,7 +4329,7 @@ export class Api<SecurityDataType> {
     * @request POST:/orgs/{org}/repos
     * @description Create a new repository for the authenticated user. OAuth users must supply. repo scope.. 
     */
-    reposCreate: (org: string, body: any, params?: RequestParams) =>
+    reposCreate: (org: string, body: postRepo, params?: RequestParams) =>
       this.request<repos>(`/orgs/${org}/repos`, "POST", params, body),
 
 
@@ -4503,7 +4503,7 @@ export class Api<SecurityDataType> {
     * @request PATCH:/repos/{owner}/{repo}/comments/{commentId}
     * @description Update a commit comment.
     */
-    commentsPartialUpdate: (owner: string, repo: string, commentId: number, body: any, params?: RequestParams) =>
+    commentsPartialUpdate: (owner: string, repo: string, commentId: number, body: commentBody, params?: RequestParams) =>
       this.request<commitComment>(`/repos/${owner}/${repo}/comments/${commentId}`, "PATCH", params, body),
 
 
@@ -4824,7 +4824,7 @@ export class Api<SecurityDataType> {
     * @request POST:/repos/{owner}/{repo}/hooks
     * @description Create a hook.
     */
-    hooksCreate: (owner: string, repo: string, body: any, params?: RequestParams) =>
+    hooksCreate: (owner: string, repo: string, body: hookBody, params?: RequestParams) =>
       this.request<hook>(`/repos/${owner}/${repo}/hooks`, "POST", params, body),
 
 
@@ -4853,7 +4853,7 @@ export class Api<SecurityDataType> {
     * @request PATCH:/repos/{owner}/{repo}/hooks/{hookId}
     * @description Edit a hook.
     */
-    hooksPartialUpdate: (owner: string, repo: string, hookId: number, body: any, params?: RequestParams) =>
+    hooksPartialUpdate: (owner: string, repo: string, hookId: number, body: hookBody, params?: RequestParams) =>
       this.request<hook>(`/repos/${owner}/${repo}/hooks/${hookId}`, "PATCH", params, body),
 
 
@@ -4880,7 +4880,7 @@ export class Api<SecurityDataType> {
     * @request POST:/repos/{owner}/{repo}/issues
     * @description Create an issue.. Any user with pull access to a repository can create an issue.. 
     */
-    issuesCreate: (owner: string, repo: string, body: any, params?: RequestParams) =>
+    issuesCreate: (owner: string, repo: string, body: issue, params?: RequestParams) =>
       this.request<issue>(`/repos/${owner}/${repo}/issues`, "POST", params, body),
 
 
@@ -4918,7 +4918,7 @@ export class Api<SecurityDataType> {
     * @request PATCH:/repos/{owner}/{repo}/issues/comments/{commentId}
     * @description Edit a comment.
     */
-    issuesCommentsPartialUpdate: (owner: string, repo: string, commentId: number, body: any, params?: RequestParams) =>
+    issuesCommentsPartialUpdate: (owner: string, repo: string, commentId: number, body: commentBody, params?: RequestParams) =>
       this.request<issuesComment>(`/repos/${owner}/${repo}/issues/comments/${commentId}`, "PATCH", params, body),
 
 
@@ -4958,7 +4958,7 @@ export class Api<SecurityDataType> {
     * @request PATCH:/repos/{owner}/{repo}/issues/{number}
     * @description Edit an issue.. Issue owners and users with push access can edit an issue.. 
     */
-    issuesPartialUpdate: (owner: string, repo: string, number: number, body: any, params?: RequestParams) =>
+    issuesPartialUpdate: (owner: string, repo: string, number: number, body: issue, params?: RequestParams) =>
       this.request<issue>(`/repos/${owner}/${repo}/issues/${number}`, "PATCH", params, body),
 
 
@@ -4978,7 +4978,7 @@ export class Api<SecurityDataType> {
     * @request POST:/repos/{owner}/{repo}/issues/{number}/comments
     * @description Create a comment.
     */
-    issuesCommentsCreate: (owner: string, repo: string, number: number, body: any, params?: RequestParams) =>
+    issuesCommentsCreate: (owner: string, repo: string, number: number, body: commentBody, params?: RequestParams) =>
       this.request<issuesComment>(`/repos/${owner}/${repo}/issues/${number}/comments`, "POST", params, body),
 
 
@@ -5016,7 +5016,7 @@ export class Api<SecurityDataType> {
     * @request POST:/repos/{owner}/{repo}/issues/{number}/labels
     * @description Add labels to an issue.
     */
-    issuesLabelsCreate: (owner: string, repo: string, number: number, body: any, params?: RequestParams) =>
+    issuesLabelsCreate: (owner: string, repo: string, number: number, body: emailsPost, params?: RequestParams) =>
       this.request<label>(`/repos/${owner}/${repo}/issues/${number}/labels`, "POST", params, body),
 
 
@@ -5025,7 +5025,7 @@ export class Api<SecurityDataType> {
     * @request PUT:/repos/{owner}/{repo}/issues/{number}/labels
     * @description Replace all labels for an issue.. Sending an empty array ([]) will remove all Labels from the Issue.. 
     */
-    issuesLabelsUpdate: (owner: string, repo: string, number: number, body: any, params?: RequestParams) =>
+    issuesLabelsUpdate: (owner: string, repo: string, number: number, body: emailsPost, params?: RequestParams) =>
       this.request<label>(`/repos/${owner}/${repo}/issues/${number}/labels`, "PUT", params, body),
 
 
@@ -5054,7 +5054,7 @@ export class Api<SecurityDataType> {
     * @request POST:/repos/{owner}/{repo}/keys
     * @description Create a key.
     */
-    keysCreate: (owner: string, repo: string, body: any, params?: RequestParams) =>
+    keysCreate: (owner: string, repo: string, body: UserKeysPost, params?: RequestParams) =>
       this.request<UserKeysKeyId>(`/repos/${owner}/${repo}/keys`, "POST", params, body),
 
 
@@ -5092,7 +5092,7 @@ export class Api<SecurityDataType> {
     * @request POST:/repos/{owner}/{repo}/labels
     * @description Create a label.
     */
-    labelsCreate: (owner: string, repo: string, body: any, params?: RequestParams) =>
+    labelsCreate: (owner: string, repo: string, body: emailsPost, params?: RequestParams) =>
       this.request<label>(`/repos/${owner}/${repo}/labels`, "POST", params, body),
 
 
@@ -5121,7 +5121,7 @@ export class Api<SecurityDataType> {
     * @request PATCH:/repos/{owner}/{repo}/labels/{name}
     * @description Update a label.
     */
-    labelsPartialUpdate: (owner: string, repo: string, name: string, body: any, params?: RequestParams) =>
+    labelsPartialUpdate: (owner: string, repo: string, name: string, body: emailsPost, params?: RequestParams) =>
       this.request<label>(`/repos/${owner}/${repo}/labels/${name}`, "PATCH", params, body),
 
 
@@ -5157,7 +5157,7 @@ export class Api<SecurityDataType> {
     * @request POST:/repos/{owner}/{repo}/milestones
     * @description Create a milestone.
     */
-    milestonesCreate: (owner: string, repo: string, body: any, params?: RequestParams) =>
+    milestonesCreate: (owner: string, repo: string, body: milestoneUpdate, params?: RequestParams) =>
       this.request<milestone>(`/repos/${owner}/${repo}/milestones`, "POST", params, body),
 
 
@@ -5186,7 +5186,7 @@ export class Api<SecurityDataType> {
     * @request PATCH:/repos/{owner}/{repo}/milestones/{number}
     * @description Update a milestone.
     */
-    milestonesPartialUpdate: (owner: string, repo: string, number: number, body: any, params?: RequestParams) =>
+    milestonesPartialUpdate: (owner: string, repo: string, number: number, body: milestoneUpdate, params?: RequestParams) =>
       this.request<milestone>(`/repos/${owner}/${repo}/milestones/${number}`, "PATCH", params, body),
 
 
@@ -5213,7 +5213,7 @@ export class Api<SecurityDataType> {
     * @request PUT:/repos/{owner}/{repo}/notifications
     * @description Mark notifications as read in a repository.. Marking all notifications in a repository as "read" removes them from the. default view on GitHub.com.. 
     */
-    notificationsUpdate: (owner: string, repo: string, body: any, params?: RequestParams) =>
+    notificationsUpdate: (owner: string, repo: string, body: notificationMarkRead, params?: RequestParams) =>
       this.request<any>(`/repos/${owner}/${repo}/notifications`, "PUT", params, body),
 
 
@@ -5269,7 +5269,7 @@ export class Api<SecurityDataType> {
     * @request PATCH:/repos/{owner}/{repo}/pulls/comments/{commentId}
     * @description Edit a comment.
     */
-    pullsCommentsPartialUpdate: (owner: string, repo: string, commentId: number, body: any, params?: RequestParams) =>
+    pullsCommentsPartialUpdate: (owner: string, repo: string, commentId: number, body: commentBody, params?: RequestParams) =>
       this.request<pullsComment>(`/repos/${owner}/${repo}/pulls/comments/${commentId}`, "PATCH", params, body),
 
 
@@ -5372,7 +5372,7 @@ export class Api<SecurityDataType> {
     * @request POST:/repos/{owner}/{repo}/releases
     * @description Create a release. Users with push access to the repository can create a release.. 
     */
-    releasesCreate: (owner: string, repo: string, body: any, params?: RequestParams) =>
+    releasesCreate: (owner: string, repo: string, body: ReleaseCreate, params?: RequestParams) =>
       this.request<release>(`/repos/${owner}/${repo}/releases`, "POST", params, body),
 
 
@@ -5428,7 +5428,7 @@ export class Api<SecurityDataType> {
     * @request PATCH:/repos/{owner}/{repo}/releases/{id}
     * @description Users with push access to the repository can edit a release
     */
-    releasesPartialUpdate: (owner: string, repo: string, id: string, body: any, params?: RequestParams) =>
+    releasesPartialUpdate: (owner: string, repo: string, id: string, body: ReleaseCreate, params?: RequestParams) =>
       this.request<release>(`/repos/${owner}/${repo}/releases/${id}`, "PATCH", params, body),
 
 
@@ -5813,7 +5813,7 @@ export class Api<SecurityDataType> {
     * @request POST:/user/emails
     * @description Add email address(es).. You can post a single email address or an array of addresses.. 
     */
-    emailsCreate: (body: any, params?: RequestParams) =>
+    emailsCreate: (body: emailsPost, params?: RequestParams) =>
       this.request<any>(`/user/emails`, "POST", params, body),
 
 
@@ -5885,7 +5885,7 @@ export class Api<SecurityDataType> {
     * @request POST:/user/keys
     * @description Create a public key.
     */
-    keysCreate: (body: any, params?: RequestParams) =>
+    keysCreate: (body: UserKeysPost, params?: RequestParams) =>
       this.request<UserKeysKeyId>(`/user/keys`, "POST", params, body),
 
 
@@ -5930,7 +5930,7 @@ export class Api<SecurityDataType> {
     * @request POST:/user/repos
     * @description Create a new repository for the authenticated user. OAuth users must supply. repo scope.. 
     */
-    reposCreate: (body: any, params?: RequestParams) =>
+    reposCreate: (body: postRepo, params?: RequestParams) =>
       this.request<repos>(`/user/repos`, "POST", params, body),