+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # next release  
 Breaking Changes:  
   - Rename default typescript api file name (prev `api.ts`, now `Api.ts`)  
+Features:  
+  - Improve response body type definitions  
 
 # 1.3.0  
 Features: